new powering method and new test

+++ src/test/java/testTravis/PoweringTest.java
@@ -6,11 +6,12 @@ import org.junit.Test;
 
 public class PoweringTest {
 
+	
 	@Test
 	public void poweringTest() {
-		int result = FunKwadratowa.powering(2,2);
+		int result = (int) FunKwadratowa.powering(2,2);
         int expectedResult = 4;
         assertEquals(result, expectedResult);
-
 	}
+
 }
